changed period year date types added reference accepted name

+++ src/base/models.py
@@ -35,8 +35,8 @@ def create_or_update_user_profile(sender, instance, created, **kwargs):
 
 class Period(models.Model):
     name = models.CharField(max_length=255)
-    year_start = models.SmallIntegerField(blank=True, null=True, verbose_name='start year')
-    year_end = models.SmallIntegerField(blank=True, null=True, verbose_name='end year')
+    year_start = models.IntegerField(blank=True, null=True, verbose_name='start year')
+    year_end = models.IntegerField(blank=True, null=True, verbose_name='end year')
 
     def __str__(self):
         start = self.year_start or ''
